smolscale: Maintain rowstride as size_t to avoid overflows
The officially supported src and dest dimensions are 65535x65535
pixels. With four channels, the allocation size can go above 4GiB
and beyond the range of a 32-bit register.

Since operations are performed incrementally on individual rows,
this will currently not cause any overflows, but we'd rather be
proactive in case we calculate full offsets at some point in the future.

+++ chafa/internal/smolscale/smolscale-private.h
@@ -11,6 +11,7 @@
 
 #undef SMOL_ENABLE_ASSERTS
 
+#include <stddef.h>
 #include <stdint.h>
 #include "smolscale.h"
 
@@ -338,8 +339,8 @@ struct SmolScaleCtx
     const char *src_pixels;
     char *dest_pixels;
 
-    uint32_t src_rowstride;
-    uint32_t dest_rowstride;
+    size_t src_rowstride;
+    size_t dest_rowstride;
 
     SmolPixelType src_pixel_type, dest_pixel_type;
     SmolStorageType storage_type;
